+++ src/main/java/school/faang/user_service/controller/event/EventParticipationController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import school.faang.user_service.dto.UserDto;
+import school.faang.user_service.dto.event.UserDto;
 import school.faang.user_service.service.event.EventParticipationService;
 
 import java.util.List;
