+++ src/main/java/school/faang/user_service/dto/event/UserDto.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.dto;
+package school.faang.user_service.dto.event;
 
 import lombok.Data;
