+++ src/main/java/school/faang/user_service/mapper/event/UserMapper.java
@@ -1,7 +1,7 @@
 package school.faang.user_service.mapper.event;
 
 import org.mapstruct.Mapper;
-import school.faang.user_service.dto.UserDto;
+import school.faang.user_service.dto.event.UserDto;
 import school.faang.user_service.entity.User;
 
 @Mapper(componentModel = "spring")
