+++ src/main/java/school/faang/user_service/service/event/EventParticipationService.java
@@ -2,7 +2,7 @@ package school.faang.user_service.service.event;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import school.faang.user_service.dto.UserDto;
+import school.faang.user_service.dto.event.UserDto;
 import school.faang.user_service.mapper.event.UserMapper;
 import school.faang.user_service.repository.event.EventParticipationRepository;
